Add blog post: Solving Data Consistency in Distributed Systems with the Transactional Outbox (#354)
* Add blog post on Transactional Outbox

Title: Solving Data Consistency in Distributed Systems with the Transactional Outbox

* Add a few words to spelling list

* Bump date on blog post

* Clarified comparison between Event Sourcing and Transactional Outbox

Improved explanation of Event Sourcing infrastructure and querying requirements, including its limitations in solving the dual-write problem for external communication. Refined rationale for choosing Transactional Outbox in project with multiple external integrations. Enhanced clarity and flow to better reflect architectural trade-offs and project constraints.

* Bump date on transactional outbox blog post

* Update Outbox blog post to use message rather than event

... this is as the Transactional Outbox Pattern is solution when
second system is an external service, not just when it is an internal
component that can respond to events.

* Add more details to lessons learnt to Outbox blog post

* Outbox blog post content clean up

* Fix typos: Outbox pattern blog post